+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `--debug` flag now dumps the internal expressions as well
 - hevm now uses the forge-std library's way of detecting failures, i.e. through
   reverting with a specific error code unless --assertion-type DSTest is passed
+- Default max iterations is 5 now. `--max-iters -1` now signals no bound. This change is to match other
+  symbolic execution frameworks' default bound and to not go into an infinite loop by default when
+  there could be other, interesting and reachable bugs in the code
 
 ## Added
 - More POr and PAnd rules
@@ -48,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Better and more pretty debug messages
 
 ## Fixed
+- `vm.prank` is now respected during calls to create
 - `concat` is a 2-ary, not an n-ary function in SMT2LIB, declare-const does not exist in QF_AUFBV, replacing
    with declare-fun
 - CVC5 needs `--incremental` flag to work properly in abstraction-refinement mode